// portal/customer-portal-vivo-angular/src/app/enumerators/address-type.enum.ts
export enum AddressType {

    // DECRIÇÃO DOS ENUMERADORES
    ENDERECO_PESSOA_FISIACA = 'Endereço pessoa física',
    ENDERECO_PESSOA_JURIDICA = 'Endereço pessoa jurídica',
    ENDERECO_INSTALACAO = 'Endereço de instalação',
    ENDERECO_CORRESPONDENCIA = 'Endereço para correspondência',

    // IDENTIFICAÇÃO DOS ENUMERADORES
    ENDERECO_PESSOA_FISIACA_ID = 'Home',
    ENDERECO_PESSOA_JURIDICA_ID = 'Comercial',
    ENDERECO_INSTALACAO_ID = 'Delivery',
    ENDERECO_CORRESPONDENCIA_ID = 'Correspondence'

}