// portal/customer-portal-vivo-angular/src/app/enumerators/solicitation-status.enum.ts
export enum SolicitationStatus {
    SOLICITATION_STATUS_OPEN = 'Open',
    SOLICITATION_STATUS_PENDING = 'Pending',
    SOLICITATION_STATUS_IN_ANALISIS = 'InAnalysis',
    SOLICITATION_STATUS_WITH_PROBLEM = 'WithProblem',
    SOLICITATION_STATUS_IN_CONFIGURATION = 'InConfiguration',
    SOLICITATION_STATUS_CLOSE = 'Close',
    SOLICITATION_STATUS_INVALID = 'Invalid',

    SOLICITATION_STATUS_OPEN_LABEL = 'Aberta',
    SOLICITATION_STATUS_PENDING_LABEL = 'Pendente',
    SOLICITATION_STATUS_IN_ANALISIS_LABEL = 'Em analise',
    SOLICITATION_STATUS_WITH_PROBLEM_LABEL = 'Com problema',
    SOLICITATION_STATUS_IN_CONFIGURATION_LABEL = 'Em configuração',
    SOLICITATION_STATUS_CLOSE_LABEL = 'Fechada',
    SOLICITATION_STATUS_INVALID_LABEL = 'Invalida'
}